statsd: Allow buffering for the clients
The new `Buffer` wrapper allow wrapping any UDP clients in a "buffering
mode" that batches incoming metrics into a buffer and eventually flushes
them to the statsd-compatible server.

The receiving server *must* support the multi-metric StatsD protocol
(namely, parsing several metrics in the same packet, separated by
newlines). Both the original StatsD and now Brubeck support this. By
sending several metrics in a single packet, we reduce the amount of
network operations (both in the client and the server) and the amount of
context switches to the kernel UDP stack, significantly increasing the
throughput of the server implementation.

The default maximum packet size has been set at 512 bytes, which is a
"safe size" to reliably transmit an atomic packet through the internet:
The MTU on the internet is 576, and the size of the IPv4 header is 20
bytes, and the UDP header 8 bytes.  This leaves 548 bytes available for
user data. We cap it at 512 to add some headspace (most other UDP-based
protocols, such as DNS, do the same thing).